update example to test multiple ad tags
